Introduce preliminary code for a UDP tracker manager.
The UDP tracker manager process is responsible for keeping an UDP
socket open to communicate with the tracker on. The idea is that
tracker_communication processes can ask this gen_server to do UDP
requests on behalf of the tracker_comm process. A message will
asynchronously arrive to the tracker_comm process when the request
finishes. It will stay silent should the tracker decide not to return
any value. Thus the tracker_comm process is still responsible for
having its own timeout scheme.

As the protocol ends up being a bit complex, it takes some more code
to get this right. Also, the code could use a cleanup-run as it turns
out it is a bit harder to write than what I first imagined.